+++ bootstrap/providers.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Modules\ActivityLog\ActivityLogServiceProvider;
+use App\Modules\Admin\AdminServiceProvider;
 use App\Modules\Auth\AuthServiceProvider;
 use App\Modules\Authorization\AuthorizationServiceProvider;
 use App\Modules\Dashboard\DashboardServiceProvider;
@@ -18,6 +19,7 @@ return [
     // Module Service Providers
     ActivityLogServiceProvider::class,
     AuthorizationServiceProvider::class,
+    AdminServiceProvider::class,
     AuthServiceProvider::class,
     LandingServiceProvider::class,
     PlanningServiceProvider::class,